+++ packages/apple-llm/src/NativeAppleSpeech.ts
@@ -1,35 +1,21 @@
 import type { TurboModule } from 'react-native'
 import { TurboModuleRegistry } from 'react-native'
 
-export interface TranscriptionSegment {
-  text: string
-  startSecond: number
-  endSecond: number
-}
-
-export interface TranscriptionResult {
-  segments: TranscriptionSegment[]
-  duration: number
+export interface SpeechOptions {
+  language?: string
+  voice?: string
 }
 
 export interface Spec extends TurboModule {
-  isAvailable(language: string): boolean
-  prepare(language: string): Promise<void>
-}
-
-declare global {
-  function __apple__llm__transcribe__(
-    data: ArrayBufferLike,
-    language: string
-  ): Promise<TranscriptionResult>
+  isAvailable(): boolean
+  generate(text: string, options?: SpeechOptions): Promise<void>
 }
 
 const NativeAppleSpeech =
   TurboModuleRegistry.getEnforcing<Spec>('NativeAppleSpeech')
 
 export default {
-  transcribe: (data: ArrayBufferLike, language: string) =>
-    globalThis.__apple__llm__transcribe__(data, language),
-  prepare: NativeAppleSpeech.prepare,
+  generate: (text: string, options?: SpeechOptions) =>
+    NativeAppleSpeech.generate(text, options),
   isAvailable: NativeAppleSpeech.isAvailable,
 }
