+++ packages/apple-llm/src/index.ts
@@ -2,4 +2,5 @@ export { apple, createAppleProvider } from './ai-sdk'
 export { default as AppleEmbeddings } from './NativeAppleEmbeddings'
 export { default as AppleFoundationModels } from './NativeAppleLLM'
 export { default as AppleSpeech } from './NativeAppleSpeech'
+export { default as AppleTranscription } from './NativeAppleTranscription'
 export { default as AppleUtils } from './NativeAppleUtils'
