+++ website/src/docs/apple/_meta.json
@@ -2,5 +2,6 @@
   { "type": "file", "name": "getting-started", "label": "Getting Started" },
   { "type": "file", "name": "generating", "label": "Generating" },
   { "type": "file", "name": "embeddings", "label": "Embeddings" },
-  { "type": "file", "name": "transcribe", "label": "Transcribe" }
+  { "type": "file", "name": "transcription", "label": "Transcription" },
+  { "type": "file", "name": "speech", "label": "Speech" }
 ]
