+++ website/src/docs/apple/speech.md
@@ -1,0 +1,3 @@
+# Speech
+
+Convert text to speech using Apple's on-device speech synthesis capabilities.
